Fix position for large nodes
The previous implementation only took into account the root of the
previous block to determine width. It's possible, however, to have a
node that is in a block that is larger than the rest. We should capture
the maximum width for each block and store it in the root. Later we can
look up the max width to properly place adjacent nodes.

This change also fixes some misuses of the sink object.

Fixes #3.