Tweaks to default attribute handling for nodes
This change ensures that we only try to apply the default attributes to
a node the first time it is created, which can happen with an explicit
node statement or implicitly when an undefined node is named in an edge
statement. We take special care to ensure that the node's label
attribute is only set if it is not already defined in the default
attribute set.