graphlib-dot: Convert most of the JavaScript to TypeScript
This reverts the conversion of dot-grammar to TypeScript, which didn't
help anything as it wasn't gathering any types from the rules. Now the
grammar's types are declared manually.

Still not sure what version of JavaScript I'm supporting. I've got a
mix of old and new features in there, and stuff that almost certainly
doesn't work on Node.js 10. Probably will update the rules here in a bit
to help.